GetServiceDtoMapperTest fixed
